Improve main loop once more *with help from claude
Actually account for lag.

I actually found the `lime.ui.Window.onClose` function does not work anymore with my new main loop but I tried basically fixing that with just a windowEvent call which didn't work. But this is just a fork of lime meant for funkin' view anyway. HEY! I found that the window_event is still being marked as an input meaning I had to fix that. Sorry for the confusion, by the way.

Oh yeah and how it handles the lag is that instead of it jittering like normal, it detects if it had lagged or not, virtually, to reset your timing base. The weirdest part? In this version, it works really well. I think that's kind of weird to be completely honest cuz I haven't done much except ai to achieve this.